Added Geometry items. 2nd

+++ extra/actions/GenerateMesh/script.js
@@ -142,7 +142,7 @@ class Script {
           items: [
             {
               name: 'index',
-              dataType: 'Uint16',
+              dataType: 'Uint32',
               count: 3,
             }
           ]
